+++ src/parser.rs
@@ -3,6 +3,7 @@ use crate::models::case::CaseConfig;
 use crate::models::variant::{VariantCategory, VariantType};
 use crate::parse::cytobands::set_cytobands;
 use crate::parse::vcf::process_vcf;
+use mongodb::bson::Document;
 use std::collections::{HashMap, HashSet};
 
 /// Parses and processes all clinical VCFs provided for a case.
@@ -16,6 +17,7 @@ use std::collections::{HashMap, HashSet};
 pub fn parse(
     config: &CaseConfig,
     gene_to_panels: &HashMap<i32, HashSet<String>>,
+    hgncid_to_gene: &HashMap<i32, Document>,
 ) -> Result<(), Box<dyn std::error::Error>> {
     let genome_build = GenomeBuild::from_str(&config.human_genome_build)
         .map_err(|_| format!("Invalid genome build: {}", config.human_genome_build))?;
@@ -47,6 +49,7 @@ pub fn parse(
                 &cytobands,
                 &config.samples,
                 gene_to_panels,
+                hgncid_to_gene,
             );
         }
     }